Safety copy of an old version of the softwares
This backup was loaded in case the delivered software fails due to problems with modules/virtual environment of changes made at the last minute. 
Obviously compared to the delivered software there is nothing conceptually different, this is simply the version of the software prior to the delivery of the thesis (there was no additional work hour).